requirements.txt
pytest
pytest-asyncio
playwright
allure-pytest
pytest-html
python-dotenv
